requirements.txt
psycopg2-binary
pymongo
kafka-python
faker
pandas
clickhouse-driver
apache-airflow==2.7.2
pyarrow>=10.0.1,<10.1.0

aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
motor==3.7.1
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
